backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1

torch==2.1.1
torchvision==0.16.1

numpy==1.26.2
Pillow==10.1.0
scikit-image==0.22.0
opencv-python-headless

python-multipart==0.0.6
slowapi==0.1.9
prometheus-client==0.20.0

pydantic-settings==2.1.0
python-dotenv==1.0.0


#naman
